Few changes.

+++ src/main/java/ua/sumdu/group73/model/OracleDataBase.java
@@ -215,7 +215,7 @@ public class OracleDataBase implements UserDBInterface, PicturesDBInterface,
 		            String eMail = rs.getString("EMAIL");
 		            String phone = rs.getString("PHONE");
 		            String status = rs.getString("STATUS");
-		            int age = rs.getByte("AGE");
+		            int age = rs.getInt("AGE");
 		            Date regDate = rs.getTimestamp("REGISTRATION_DATE");
 		            
 		            if (rs.getString("PASSWORD").equals(password) &&
@@ -247,7 +247,7 @@ public class OracleDataBase implements UserDBInterface, PicturesDBInterface,
 		            String login = rs.getString("LOGIN");
 		            String phone = rs.getString("PHONE");
 		            String status = rs.getString("STATUS");
-		            int age = rs.getByte("AGE");
+		            int age = rs.getInt("AGE");
 		            Date regDate = rs.getTimestamp("REGISTRATION_DATE");
 		            
 		            if (rs.getString("PASSWORD").equals(password) &&
@@ -358,7 +358,7 @@ public class OracleDataBase implements UserDBInterface, PicturesDBInterface,
 		            String password = rs.getString("PASSWORD");
 		            String name = rs.getString("NAME");
 		            String secondName = rs.getString("SECOND_NAME");
-		            int age = rs.getByte("AGE");
+		            int age = rs.getInt("AGE");
 		            String eMail = rs.getString("EMAIL");
 		            String phone = rs.getString("PHONE");
 		            String status = rs.getString("STATUS");
